import from anywhere.

+++ rye.py
@@ -12,6 +12,23 @@ PROFILE = os.getenv("RYE_PROFILE")
 
 PATH_MATCH = re.compile('(.*)/src/(.*)').match
 
+def TranslateModuleAndDependencies(filename, longmod, mod, cwd, twd, did):
+  imports = TranslateModule(filename, longmod, mod, cwd)
+  did[longmod] = True
+
+  for k, v in imports.items():
+    # print >> sys.stderr, "####### %s -> %s" % (k, vars(v))
+    if v.imported[0] == 'go':
+      continue # Don't traverse "go" dependencies.
+
+    longpath2 = v.imported
+    longmod2 = '/'.join(longpath2)
+    mod2 = longpath2[-1]
+    filename2 = '/' + twd + '/src/' + longmod2 + '.py'
+
+    if not did.get(longmod2):
+      TranslateModuleAndDependencies(filename2, longmod2, mod2, os.path.dirname(longmod2), twd, did)
+
 def TranslateModule(filename, longmod, mod, cwp):
   print >>sys.stderr, '*** TranslateModule', [filename, longmod, mod, cwp]
   d = os.path.dirname(filename)
@@ -25,35 +42,33 @@ def TranslateModule(filename, longmod, mod, cwp):
   wpath = os.path.join(d, b, 'ryemodule.go') 
 
   # BUG: If we don't recompile one, we may not notice its dirty dependency.
-  #try:
-  #  w_st = os.stat(wpath)
-  #  w_mtime = w_st.st_mtime
-  #except:
-  #  w_mtime = 0
-  #r_st = os.stat(filename)
-  #r_mtime = r_st.st_mtime
-  #if w_mtime > r_mtime:
-  #  print >> sys.stderr, "*** ALREADY COMPILED: %s" % filename
-  #  return {}
+  try:
+    w_st = os.stat(wpath)
+    w_mtime = w_st.st_mtime
+  except:
+    w_mtime = 0
+  r_st = os.stat(filename)
+  r_mtime = r_st.st_mtime
+  already_compiled = (w_mtime > r_mtime)
 
-  sys.stdout = open(wpath, 'w')
   program = open(filename).read()
   words = tr.Lex(program).tokens
-
-  #print >> sys.stderr, "\n\n(TOKENS)<<<", words, "\n\n"
   words = list(tr.SimplifyContinuedLines(words))
-  #print >> sys.stderr, "\n\n(TOKENS)>>>", words, "\n\n"
-
-  parser = tr.Parser(program, words, -1)
+  parser = tr.Parser(program, words, -1, cwp)
   try:
     tree = parser.Csuite()
   except:
-    print >> sys.stderr, "\n*** ERROR: ", sys.exc_info()[1]
-    print >> sys.stderr, "\n*** OCCURRED BEFORE THIS: ", parser.Rest()[:200], '......'
     print >> sys.stderr, "\n*** TRACEBACK:"
     traceback.print_tb(sys.exc_info()[2])
+    print >> sys.stderr, "\n*** OCCURRED BEFORE THIS: ", repr(parser.Rest()[:100])
+    print >> sys.stderr, "\n*** ERROR: ", sys.exc_info()[1]
     sys.exit(13)
 
+  if already_compiled:
+    print >> sys.stderr, "Already Compiled:", longmod
+    sys.stdout = open('/dev/null', 'w')
+  else:
+    sys.stdout = open(wpath, 'w')
   gen = tr.CodeGen(None)
   gen.GenModule(mod, longmod, tree, cwp)
   sys.stdout.close()
@@ -114,7 +129,7 @@ def LongMod(*TODO):
     return '%s/%s/%s' % (cwd, d, mod)
 
 def BuildRun(to_run, args):
-  print >>sys.stderr, "*** BUILD", to_run, args
+  print >>sys.stderr, "*** BUILD", args
   pwd = os.getcwd()
   m = PATH_MATCH(pwd)
   if not m:
@@ -128,10 +143,10 @@ def BuildRun(to_run, args):
   first = True
   did = {}
   full_run_args = []
-  todo = [ args ]
-  while todo:
+  yet = [ args ]  # List of args lists.
+  while yet:
     run_args = None
-    chunk = todo.pop(0)
+    chunk = yet.pop(0)
     for a in chunk:
       if run_args is not None:  # After --, just collect run_args.
         run_args.append(a)
@@ -141,19 +156,25 @@ def BuildRun(to_run, args):
         run_args = []
         continue
 
-      if did.get(a):  # Don't do any twice.
-        continue
-
       d = os.path.dirname(a)
-
       mod = os.path.basename(a).split('.')[0]  # Part before '.' in basename becomes package name.
 
       if d == '.' or d == "":
         longmod = '%s/%s' % (cwd, mod)
       else:
         longmod = '%s/%s/%s' % (cwd, d, mod)
-      longdir = os.path.dirname(longmod)
+      longmod = '/'.join(tr.CleanPath('/', longmod))
+      if did.get(longmod):  # Don't do any twice.
+        continue
 
+      if first:
+        main_longmod = longmod
+        main_mod = mod
+        main_filename = WriteMain(a, longmod, mod)
+        first = False
+
+      imports = TranslateModuleAndDependencies(a, longmod, mod, cwd, twd, did)
+      """
       imports = TranslateModule(a, longmod, mod, cwd)
       did[a] = True
 
@@ -164,18 +185,19 @@ def BuildRun(to_run, args):
         first = False
 
       for k, v in imports.items():
-        # print >> sys.stderr, "####### %s -> %s" % (k, vars(v))
-        if v.fromWhere is None:  # Todo, handle more fromWhere.
+        print >> sys.stderr, "####### %s -> %s" % (k, vars(v))
+        if v.imported[0] != 'go':
 
-          if v.imported[:len(cwd_split)] != cwd_split:
-            raise Exception("Cannot handle this import yet: %s (vs %s)", v.imported, cwd_split)
+          #I# if v.imported[:len(cwd_split)] != cwd_split:
+          #I#   raise Exception("Cannot handle this import yet: %s (vs %s)", v.imported, cwd_split)
 
           impfile = '%s.py' % ('/'.join(v.imported[len(cwd_split):]))
           # print >> sys.stderr, "IMPFILE %s" % impfile
 
           if not did.get(impfile):
-            todo.append([impfile])
+            yet.append([impfile])
             # print >> sys.stderr, "ADDED %s" % impfile
+      """
 
     full_run_args += (run_args if run_args else [])
 
@@ -185,7 +207,7 @@ def BuildRun(to_run, args):
 
   cmd = "set -x; go build -o '%s' '%s'" % (target, main_filename)
   cmd = ['go', 'build', '-o', target, main_filename]
-  print >> sys.stderr, "+ %s" % repr(cmd)
+  print >> sys.stderr, "+ %s" % ' '.join(["'%s'" % s for s in cmd])
   status = Execute(cmd)
   if status:
     print >> sys.stderr, "%s: Exited with status %d" % (main_longmod, status)
